// src/config/constants/types.ts
export interface Address {
  1?: string
  4?: string
  97?: string
  56?: string
}

export interface Token {
  symbol: string
  address?: Address
  decimals?: number
  projectLink?: string
  busdPrice?: string
}

export interface ToggleMenuItem {
  chainID: number
  key: string
  value: string
}

export interface Symbols {
  1?: string
  4?: string
  97?: string
  56?: string
}

export interface Tokens {
  1?: Token
  4?: Token
  97?: Token
  56?: Token
  busdPrice?: number
}
// ...
export interface FarmConfig {
  pid: number
  lpSymbol: Symbols
  lpAddresses: Address
  token: Tokens
  quoteToken: Tokens
  multiplier?: string
  isCommunity?: boolean
  dual?: {
    rewardPerBlock: number
    earnLabel: string
    endBlock: number
  }
}